Abstract Dock from Refinery
